refactoring podcast header and episode list
